% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph.pb.dat.setup.R
\name{coxph.pb.dat.setup}
\alias{coxph.pb.dat.setup}
\title{Basic Data Setup Prior to Computing the Minus Log Exact Partial Likelihood}
\usage{
coxph.pb.dat.setup(cox.fit, bl.fit)
}
\arguments{
\item{cox.fit}{fitted Cox model by \code{coxph} from the \code{survival} package.}

\item{bl.fit}{fitted baseline hazard of \code{cox.fit} by \code{survfit} from the \code{survival} package.}
}
\value{
The returned list contains the following components:

\code{time}: a vector of event times.

\code{nevent}: a vector representing the number of events at each event time.

\code{nrisk}: a vector representing the number of units at risk at each event time.

\code{failmat}: a matrix of event indicators, where rows represent subjects and columns represent event times.

\code{riskmat}: a matrix of at-risk indicators, where rows represent subjects and columns represent event times.

\code{xmat}: a covariate matrix, where rows represent subjects and columns represent covariates.

\code{beta0}: the fitted slope from the initial Cox model.

\code{lambda0}: the fitted baseline hazard from the initial Cox model at each event time.
}
\description{
The exact partial likelihood involves a complex function of conditional probabilities based on the risk and event sets. This function provides essential data setup, such as distinct ordered event times, the count of individuals at risk and events at each time, risk and event sets, and the covariate matrix. These components serve as the groundwork for calculating the negative log of the exact partial likelihood. Since the exact partial likelihood depends on an unknown true baseline hazard, we substitute it with an initial Cox model, like Breslow or Efron, using its fitted baseline hazard. The output also includes the fitted slope and the baseline hazard from this initial Cox model.
}
\examples{
library(KMsurv)
data("larynx")
cox.fit=coxph(Surv(time, delta)~age+diagyr+factor(stage),data=larynx,method="efron")
bl.fit=survfit(cox.fit, newdata=data.frame(t(c(age=0,diagyr=0,stage=1))))
pb.data.setup=coxph.pb.dat.setup(cox.fit,bl.fit)
print(pb.data.setup)


}
