% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph.PB.R
\name{coxph.PB}
\alias{coxph.PB}
\title{Fitting Cox Model with Exact Partial Likelihood.}
\usage{
coxph.PB(
  formula,
  data,
  baseline.data,
  bhf.initial = "efron",
  info.option = "pbd"
)
}
\arguments{
\item{formula}{a formula for the Cox model, following the same syntax as \code{coxph} from the \code{survival} package.}

\item{data}{the dataset for the Cox model, used in the same way as in \code{coxph} from the \code{survival} package.}

\item{baseline.data}{baseline data used to calculate the fitted baseline hazard of the initial Cox model. For a factor covariate, use the reference factor level; for other covariates, use zero. See the example code below for a detailed usage example.}

\item{bhf.initial}{tie correction method for the initial Cox model. The default is "efron"; other options align with the \code{method} argument in \code{coxph} from the \code{survival} package.}

\item{info.option}{method for computing the covariance estimator for the slope. The default option, "pbd", uses the inverse of the exact information matrix, computed numerically from the exact partial likelihood and evaluated at the PBD estimated slope. In contrast, the "breslow" option uses the inverse of the Breslow information, also evaluated at the PBD estimated slope. This choice influences both the covariance estimation and the Wald statistic.}
}
\value{
The returned list contains the following components:

\code{coef} fitted slope optimizing exact partial likelihood.

\code{vcov} covariance estimator for the slope.

\code{lambda0} fitted baseline hazard, optimized based on exact likelihood.

\code{logPL} vector of exact log partial likelihood evaluated at the 0 vector and \code{coef}.

\code{lrt} likelihood ratio test statistic for the entire slope, including its degrees of freedom and p-value.

\code{wald.all} wald statistic for the entire slope, including its degrees of freedom and p-value.

\code{wald.each} wald statistic and p-value for each slope.

\code{bhf.initial} the given \code{bhf.initial}.

\code{info.option} the given \code{info.option}.

\code{n} sample size of the given \code{data}.

\code{nevent} total number of events in the given \code{data}.

\code{dat} a data object created by \code{coxph.pb.dat.setup} with the initial Cox model.

\code{formula} the given \code{formula}.

\code{call} the call used to run the function.
}
\description{
This function fits a Cox model by optimizing the exact partial likelihood for the slope and the exact likelihood for the baseline hazard. It returns estimates for both the slope and baseline hazard, along with covariance estimates for the slope. Additionally, the function provides Wald and Likelihood Ratio Test (LRT) statistics and their p-values for inference. Refer to the manuscript for detailed calculations of the estimators and test statistics.
}
\examples{
library(poibin)
library(survival)
library(numDeriv)
library(KMsurv)
data("larynx")
fit_pbd=coxph.PB(Surv(time, delta)~age+diagyr+factor(stage),
                data=larynx,
                baseline.data=t(c(age=0,diagyr=0,stage=1)),
                bhf.initial="efron",info.option="pbd")
print(fit_pbd)

}
