% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infob.R
\name{infob}
\alias{infob}
\title{Calculate the Breslow Information Matrix.}
\usage{
infob(dat, pars)
}
\arguments{
\item{dat}{the basic data setup result by \code{coxph.pb.dat.setup}, where \code{nevent}, \code{riskmat}, and \code{xmat} in it are used in Breslow information calculation.}

\item{pars}{given slope.}
}
\value{
The returned value is the Breslow information matrix at a given slope.
}
\description{
This function calculates the Breslow information matrix for a given slope.
}
\examples{
library(poibin)
library(survival)
library(numDeriv)
library(KMsurv)
data("larynx")
cox.fit=coxph(Surv(time, delta)~age+diagyr+factor(stage),data=larynx,method="efron")
bl.fit=survfit(cox.fit, newdata=data.frame(t(c(age=0,diagyr=0,stage=1))))
pb.data.setup=coxph.pb.dat.setup(cox.fit,bl.fit)
br.info=infob(pb.data.setup,rep(0,5))
print(br.info)

}
