% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda0.update.R
\name{lambda0.update}
\alias{lambda0.update}
\title{Estimate the Baseline Hazard Minimizing the Negative Log Exact Likelihood.}
\usage{
lambda0.update(dat)
}
\arguments{
\item{dat}{a data object created by \code{coxph.pb.dat.setup}. In this object, the \code{beta0} represents the PBD estimated slope, rather than an initial Cox model slope. To substitute the initial slope with the PBD estimated slope, refer to the example code provided below.}
}
\value{
The returned value is the estimated baseline hazard.
}
\description{
This function estimate the baseline hazard by minimizing the negative log of the exact likelihood for the baseline hazard. Since the true slope is unknown, it is substituted with the PBD estimated slope.
}
\examples{
library(poibin)
library(survival)
library(numDeriv)
library(KMsurv)
data("larynx")
cox.fit=coxph(Surv(time, delta)~age+diagyr+factor(stage),data=larynx,method="efron")
bl.fit=survfit(cox.fit, newdata=data.frame(t(c(age=0,diagyr=0,stage=1))))
pb.data.setup=coxph.pb.dat.setup(cox.fit,bl.fit)
pb.slope.fit=lifetime.mle(dat=pb.data.setup, minusloglik=minus.log.partial.PB.discretized, starts=pb.data.setup$beta0)
pb.data.setup$beta0=as.vector(pb.slope.fit$coef)
pb.baseline.fit=lambda0.update(dat=pb.data.setup)
print(pb.baseline.fit)

}
