% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minus.log.partial.PB.discretized.R
\name{minus.log.partial.PB.discretized}
\alias{minus.log.partial.PB.discretized}
\title{Calculate the Minus Log Exact Partial Likelihood.}
\usage{
minus.log.partial.PB.discretized(dat, pars)
}
\arguments{
\item{dat}{the basic data setup result by \code{coxph.pb.dat.setup}, where the \code{lambda0}: fitted baseline hazard from the initial Cox model in it, substitutes the unknown true baseline hazard in the minus log exact partial likelihood.}

\item{pars}{given slope.}
}
\value{
The returned value is minus log partial likelihood at a given slope.
}
\description{
This function calculates the negative log of the exact partial likelihood for a given slope. Since the true baseline hazard is unknown, it is substituted with a fitted baseline hazard from an initial Cox model.
}
\examples{
library(poibin)
library(survival)
library(numDeriv)
library(KMsurv)
data("larynx")
cox.fit=coxph(Surv(time, delta)~age+diagyr+factor(stage),data=larynx,method="efron")
bl.fit=survfit(cox.fit, newdata=data.frame(t(c(age=0,diagyr=0,stage=1))))
pb.data.setup=coxph.pb.dat.setup(cox.fit,bl.fit)
minus.log.PL=minus.log.partial.PB.discretized(pb.data.setup,rep(0,5))
print(minus.log.PL)

}
